Refactor create_settings_file module location.
Moved create_settings_file to the utilities directory for better organization. Updated imports accordingly to reflect the new location. Added address_groups to the module imports for improved object handling.
