+++ scm_config_clone/__init__.py
@@ -1,5 +1,6 @@
 # scm_config_clone/__init__.py
 
-from .commands.create_settings_file import create_settings
 from .commands.objects.address import addresses
+from .commands.objects.address_groups import address_groups
 from .commands.objects.tag import tags
+from .commands.utilities.create_settings_file import create_settings
