+++ scm_config_clone/commands/utilities/create_settings_file.py
